chore(implement-feedback): - Debug Heroku and Travis build failures - Rename components that were not correctly referenced which resulted in build failures
