+++ .babelrc
@@ -1,4 +1,4 @@
 {
-  "presets": ["env", "react", "es2015"],
+  "presets": ["env", "react"],
   "plugins": ["transform-object-rest-spread"]
 }
