+++ app/client/components/client-frame/Header.jsx
@@ -7,7 +7,7 @@ import toastr from 'toastr';
 import $ from 'jquery';
 import ModalFrame from '../modal/ModalFrame';
 import GroupMembersTable from '../tables/GroupMembersTable';
-import { logout } from '../../actions/signIn';
+import { logout } from '../../actions/signIn.js';
 import setSelectedGroup from '../../actions/setSelectedGroup';
 import { setGroupMessages } from '../../actions/groupMessages';
 import { submitNewUser } from '../../actions/groupMembers';
